refactor: remove duplicate function names
