+++ src/scripts/get-cars.ts
